pyproject.toml
[tool.poetry]
name = "actionator"
version = "0.1.0"
description = ""
authors = ["obwan02 <obwan02@hotmail.co.nz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = {version = "^0.24.0.post1", extras = ["standard"]}
python-dotenv = "^1.0.0"
starlette = "^0.34.0"
pydantic = "^2.5.2"
beautifulsoup4 = "^4.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "actionator:main"
